% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rsqlite_tmp.R
\name{create_rsqlite_tmp}
\alias{create_rsqlite_tmp}
\title{Create temporary directory for RSQLite}
\usage{
create_rsqlite_tmp(
  root_dir = getwd(),
  rsqlite_tmp_folder = "rsqlite_tmp",
  run = TRUE
)
}
\arguments{
\item{root_dir}{path to create temporary directory.}

\item{rsqlite_tmp_folder}{name of temporary folder.}

\item{run}{boolean specifying whether to create temporary folder}
}
\value{
NULL.
}
\description{
Create temporary directory for RSQLite
}
\examples{
create_rsqlite_tmp()
}
